Add readonly import to code sample
Not completely sure the rules for when to have the vue imports in code samples, but not having this one threw me because the ones directly above do have the imports, and it seems each vue function is imported at least once in the example code.

I spent some time checking if readonly was a native javascript feature of proxies since I've never actually worked with them. Showing the import should quickly clear up where it's coming form.