fix, minor grammar. first-class citizen
I'd think that 'It has a [...] support.' does not sound right.
So either; 1) 'It has a first-class citizens' support', or 2) 'It has first-class citizen support' may fit better?
Option 2) technically means that first-class citizens support it. So I guess option 1) is the way to go. 
I'm not a native english speaker, so forgive me for wasting your time if I'm wrong altogether.
Cheers!