Update v3 directive example to have quoted string
I've seen this cause confusion several times. The examples don't show a variable `yellow` defined, and because a string is expected the example, as is, will fail. I believe adding the extra `'` quotes will make it clearer, and for anyone copy-pasting won't cause issues. Alternatively, the variable could be shown, but that would make examples longer.