fix: exclude style example blocks from banner styles
If the Vue School banner is open, the style example blocks on the Style
Guide will have too much top margin and padding. Detail blocks directly
above a style example block will also be unexpandable.

This is because the styles in BannerTop.vue mistakenly apply top margin
and padding to the header in style example blocks. This causes the
blocks to be too tall, as well as oveflow their container, overlaying
the detail block above and blocking clicks to the <detail> element.

Let's fix the styles for headers in BannerTop.vue to exclude those in
style example blocks.