Fix grammatical errors and introduce app.component() in introduction.md  (#1388)
* Fixed grammatical error in theme-data.js

In the Flatlogic description, the phrase "how real applications is built" is grammatically incorrect since "applications" is a plural noun but "is" refers to a singular noun. I changed it to "how real applications are built."

In the PrimeVue description, the phrase "look&feel" (which lacked spaces between the words and punctuation) has been changed to "look & feel".

* Update theme-data.js

In the Flatlogic description, "Check the admin dashboard templates" has been changed to "Check out the admin dashboard templates."

* Fix grammar and introduce the component method

* Add missing 'a'

* Update introduction to include reviewer's changes

Removed all use of the app.component() method from the introduction so as to not confuse readers.