Subject-verb agreement with "component state"
First, I know more about grammar than reactivity/components and this is my first contribution to this, but it seems "component state" is a singular object (grammatically speaking), so "is" would is the correct for of the verb. Furthermore, I think a more succinct sentence would be "Component state is composed of reactive JavaScript objects." Without using the word "composed" here the sentence seems to be saying that the two things are synonymous, which would be, strictly speaking, false.