clarify restrictions on where composables can be invoked
I have seen some vue developers interpret this section as: when calling composables in a `<script setup>` or `setup()` context, you must call them synchronously. This is good, however feel like we may need to be clear that composables also should only be called in `<script setup>` or `setup()` regardless of whether they are a/synchronously called, and for the reasons described later in the section.

Composables should, in all cases, be called:

  1. in `<script setup>` or `setup()` hook
  2. synchronously