Change <p> tag to <h2> tag
Change <p> tag to <h2> tag - the name of the ref() in this tutorial is also "p" and using a tag and a variable with the same name can be little confusing. I thought of changing the variable name from "p" to "q", for example, but that would involve making changes to more files in this tutorial.