docs(pinia): use stronger language
Use a stronger language to sell Pinia's type inference.

The word 'solid' has a neutral-positive connotation: calling a feature 'solid' implies it might is not the best feature in the system. For example, saying I have 'solid social skills' implies I have other skills that are better, like making substantial contributions to the open-source community by making PRs about language choices in documentation.

On the other hand, the word 'robust' is strictly positive: a system that is robust is implied to be complete and even battle-tested.

I believe our sales language should not only reflect the confidence we have in our products, but also guide it. That's how we'll make truly robust systems.